+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 #### 功能构建
 
+- 新建 Controller。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.system.DeriveController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.system.LoginStateController。
+
 - 为 Controller 添加新方法。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.system.LoginController。